Replaced Madgwick with gyro integration

+++ flight-code/Core/Inc/main.h
@@ -132,7 +132,7 @@ void Error_Handler(void);
 
 /* USER CODE BEGIN EFP */
 
-uint32_t micros(void);
+uint64_t micros(void);
 
 /* USER CODE END EFP */
